refactor: make the http handling code shared between iroh-gateway and iroh-one
