+++ iroh-one/src/lib.rs
@@ -1,4 +1,3 @@
-mod client;
 pub mod config;
 pub mod core;
 mod error;
